# src/main/resources/api-specs/product-api-spec.yml
openapi: 3.0.0
info:
  title: Product API
  description: API for managing marketplace products
  version: 1.0.0
paths:
  /api/v1/products:
    post:
      summary: Create a new product
      tags:
        - Products

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'

      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    get:
      summary: Get all products
      tags:
        - Products

      responses:
        '200':
          description: List of all products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'


  /api/v1/products/{id}:
    get:
      summary: Get product by id
      tags:
        - Products
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the product to retrieve
          schema:
            type: integer
            format: int64

      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'


    put:
      summary: Update product
      tags:
        - Products
      parameters:
        - in: path
          name: id
          required: true
          description: Product identifier
          schema:
            type: integer
            format: int64

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'

      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'

        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'


    delete:
      summary: Delete product
      tags:
        - Products
      parameters:
        - in: path
          name: id
          required: true
          description: Product identifier
          schema:
            type: integer
            format: int64

      responses:
        '204':
          description: Product deleted successfully

        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'




components:
  schemas:
    ProductDTO:
      type: object

      required:
        - name
        - price

      properties:
        id:
          type: integer
          format: int64
          readOnly: true

        name:
          type: string
          maxLength: 256
          minLength: 1
          description: |
            Product name.
            Must contain at least one cosmic word:
            star, galaxy, or comet.
          example: Galaxy Cat

        price:
          type: number
          minimum: 0
          example: 199.99

    ProblemDetail:
      type: object

      properties:
        type:
          type: string
          format: uri
          example: about:blank

        title:
          type: string
          example: Product not found

        status:
          type: integer
          format: int32
          example: 404

        detail:
          type: string
          example: Product with id 10 not found

        instance:
          type: string
          format: uri
          example: /api/v1/products/10